+++ src/main/java/pe/com/tatapizzeria/repository/PreciosVariantesRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.PreciosVariantesEntity;
 import java.util.List;
 
-public interface PreciosVariantesRepository extends JpaRepository<PreciosVariantesEntity, Long> {
+public interface PreciosVariantesRepository extends JpaRepository<PreciosVariantesEntity, Integer> {
 
     @Query("SELECT s FROM PreciosVariantesEntity s WHERE s.estado = true")
     List<PreciosVariantesEntity> findAllCustom();
