+++ src/main/java/pe/com/tatapizzeria/repository/ProductosRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.ProductosEntity;
 import java.util.List;
 
-public interface ProductosRepository extends JpaRepository<ProductosEntity, Long> {
+public interface ProductosRepository extends JpaRepository<ProductosEntity, Integer> {
 
     @Query("SELECT s FROM ProductosEntity s WHERE s.estado = true")
     List<ProductosEntity> findAllCustom();
