+++ src/main/java/pe/com/tatapizzeria/repository/TamanosRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.Query;
 import pe.com.tatapizzeria.entity.TamanosEntity;
 import java.util.List;
 
-public interface TamanosRepository extends JpaRepository<TamanosEntity, Long> {
+public interface TamanosRepository extends JpaRepository<TamanosEntity, Integer> {
 
     @Query("SELECT s FROM TamanosEntity s WHERE s.estado = true")
     List<TamanosEntity> findAllCustom();
